scripts: mk_modules: use make_ext4fs if exist
This patch supports make_ext4fs which is android ext4fs tools if
it exist. You don't need sudo permission if you can use the tool.

+++ scripts/mk_modules.sh
@@ -10,7 +10,11 @@ MOD_DIR="usr/tmp_mod"
 MOD_IMG="usr/modules.img"
 MOD_SIZE=32
 
-sudo ls > /dev/null
+[ -e /usr/bin/make_ext4fs ] && USE_MAKE_EXT4FS=1
+
+if [ "$USE_MAKE_EXT4FS" != "1" ]; then
+	sudo ls > /dev/null
+fi
 
 make ARCH=arm modules_prepare
 make ARCH=arm modules -j ${NCPUS}
@@ -23,15 +27,19 @@ fi
 [ -d ${MOD_DIR} ] || mkdir ${MOD_DIR}
 make ARCH=arm modules_install INSTALL_MOD_PATH=${MOD_DIR} INSTALL_MOD_STRIP=1
 
-dd if=/dev/zero of=${MOD_IMG} bs=1M count=${MOD_SIZE}
+if [ "$USE_MAKE_EXT4FS" == "1" ]; then
+	/usr/bin/make_ext4fs -b 4096 -L modules -l ${MOD_SIZE}M $MOD_IMG ${MOD_DIR}/lib/modules/
+else
+	dd if=/dev/zero of=${MOD_IMG} bs=1M count=${MOD_SIZE}
 
-mkfs.ext4 -F -b 4096 -L modules ${MOD_IMG}
+	mkfs.ext4 -F -b 4096 -L modules ${MOD_IMG}
 
-[ -d ${MOD_DIR}/mnt ] || mkdir ${MOD_DIR}/mnt
-sudo mount -o loop ${MOD_IMG} ${MOD_DIR}/mnt
-sudo cp -rf ${MOD_DIR}/lib/modules/* ${MOD_DIR}/mnt
-sync
-sudo umount ${MOD_DIR}/mnt
+	[ -d ${MOD_DIR}/mnt ] || mkdir ${MOD_DIR}/mnt
+	sudo mount -o loop ${MOD_IMG} ${MOD_DIR}/mnt
+	sudo cp -rf ${MOD_DIR}/lib/modules/* ${MOD_DIR}/mnt
+	sync
+	sudo umount ${MOD_DIR}/mnt
+fi
 
 rm -rf ${MOD_DIR}
